+++ tests/c/trajectory.c
@@ -40,8 +40,8 @@ static void test_read() {
     assert(!chfl_frame_atoms_count(frame, &natoms));
     assert(natoms == 297);
 
-    chfl_vector_t positions_0 = {0.417219f, 8.303366f, 11.737172f};
-    chfl_vector_t positions_124 = {5.099554f, -0.045104f, 14.153846f};
+    chfl_vector_t positions_0 = {0.417219, 8.303366, 11.737172};
+    chfl_vector_t positions_124 = {5.099554, -0.045104, 14.153846};
     chfl_vector_t* positions = NULL;
 
     // Check for the error when requesting non-existent velocities
@@ -87,8 +87,8 @@ static void test_read() {
     assert(c == 30.0);
     assert(!chfl_cell_free(cell));
 
-    positions_0[0] = 0.761277f;  positions_0[1] = 8.106125f;   positions_0[2] = 10.622949f;
-    positions_124[0] = 5.13242f; positions_124[1] = 0.079862f; positions_124[2] = 14.194161f;
+    positions_0[0] = 0.761277;  positions_0[1] = 8.106125;   positions_0[2] = 10.622949;
+    positions_124[0] = 5.13242; positions_124[1] = 0.079862; positions_124[2] = 14.194161;
 
     assert(!chfl_frame_positions(frame, &positions, &natoms));
     assert(natoms == 297);
@@ -198,7 +198,7 @@ static void test_write() {
 
     for (unsigned i=0; i<4; i++) {
         for (unsigned j=0; j<3; j++) {
-            positions[i][j] = (float)j + 1.0f;
+            positions[i][j] = j + 1.0;
         }
     }
 
@@ -213,7 +213,7 @@ static void test_write() {
     assert(natoms == 6);
     for (unsigned i=0; i<6; i++) {
         for (unsigned j=0; j<3; j++) {
-            positions[i][j] = (float)j + 4.0f;
+            positions[i][j] = j + 4.0;
         }
     }
     assert(!chfl_topology_append(top, atom));
